requirements.txt
tushare>=1.4.0
pandas>=2.0.0
numpy>=1.24.0
httpx>=0.27.0
openai>=1.30.0
pyyaml>=6.0
python-dotenv>=1.0.0
easytrader>=0.22.0
schedule>=1.2.0
loguru>=0.7.0
qiniu>=7.13.0
